+++ src/greentest/test__all__.py
@@ -1,6 +1,6 @@
 """Check __all__, __implements__, __extensions__, __imports__ of the modules"""
 from __future__ import print_function
-import six
+import _six as six
 import sys
 import unittest
 import types