+++ src/greentest/test__event.py
@@ -1,7 +1,7 @@
 import greentest
 import gevent
 from gevent.event import Event, AsyncResult
-from six import xrange
+from _six import xrange
 
 DELAY = 0.01