+++ src/greentest/test__order.py
@@ -1,6 +1,6 @@
 import gevent
 import greentest
-from six import xrange
+from _six import xrange
 
 
 class appender(object):