+++ src/greentest/test__socket_dns.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-import six
+import _six as six
 import re
 import greentest
 import socket
@@ -8,7 +8,7 @@ from time import time
 import gevent
 import gevent.socket as gevent_socket
 from util import log
-from six import xrange
+from _six import xrange
 
 
 resolver = gevent.get_hub().resolver