+++ src/greentest/test_hub_join_timeout.py
@@ -2,7 +2,7 @@ from contextlib import contextmanager
 import gevent
 from gevent.event import Event
 from time import time
-from six import xrange
+from _six import xrange
 
 
 SMALL = 0.1