+++ src/greentest/test_threading_2.py
@@ -1,6 +1,6 @@
 # testing gevent's Event, Lock, RLock, Semaphore, BoundedSemaphore with standard test_threading
 from __future__ import print_function
-from six import xrange
+from _six import xrange
 
 setup_ = '''from gevent import monkey; monkey.patch_all()
 from gevent.event import Event